feat(client): Move current statisticsPage (charts) state to custom hook
